Add format info to database, and accessories file
The extra field in database.ini denotes whether or not a truck uses the
new accessory-based format for paintjobs. The only trucks I know of that
do are the 2016 Scania S and R. Also added accessories.ini, to act as
an additional database defining the accessories each truck can take.

No need to update reference.txt, as the previously unused
new_truck_format variable was already described.