Add new_truck_format data to newly created trucks
When adding a truck to a truck list, or changing the supported truck for
the manual paintjob, the list's .ini file is now pre-filled with a
default accessory_name and auto-assigns all available accessories to it.

While generating mod files with accessories (i.e. new_truck_format) is
functional, managing trucks with them is not. This is the first step
towards making that a reality, followed by editing accessory_names and
assigning accessory_types to them.

Additionally, the "View/edit current truck list parameters" option on
the main menu has been re-worded to be more in line with what the end
user experiences. Editing "paintjob pack parameters" makes a bit more
sense than editing "truck list parameters".