Add generation of full paintjob packs
auto.py has been rewritten to function similarly to man.py, but for an
entire pack. A number of files were added to the ats_support_test list
as part of testing. Additionally while testing, I found that
vehicle_type had to be added to each truck list's .ini file.