Remove list_name variable
list_name has been removed, and instead moved to truck_list. Each truck
list's internal and user-facing names are now one and the same, making
things a little more simple. Most of the changes for this commit were
single-line edits, but some functions (such as rename_truck_list) had a
bit more work.

This was mostly done so that when users place their input files within
the truck list's folder, they're not exposed to the ugly internal name,
but rather the nice version.

There was also no reason to impose strict internal name limits in the
first place, as the paintjob-packer's lists have nothing to do with the
final files the game has to deal with. The lists were just limited to
match the trucks' internal_names.