Fix crash wehn renaming a truck the same name
Previously, when renaming (that is, changing the internal_name) a truck
in a list, if the user enters the same name as the truck currently has,
the program accepts it as okay, as to do the opposite would be to throw
an error and require another try. Unfortunately, it then tries to add a
section to the .ini file with the "new" name. Being identical to the
existing one, ConfigParser (which requires unique section names) throws
a fit. This has been fixed by adding None type to a boolean variable,
essentially making it a three-state boolean (this is bad practice,
shame on me).