Remove pack_name variable
The pack_name variable can just be replaced by the truck_list name, as
it is now a nice-looking name (as of d16158d). The pack_name variable is
still used in paintjob.py for two reasons; so that manual paintjobs are
not just called "manual", and so that when single.py is implemented in
the future it can work correctly.

Also fixed a bug involving commas in truck_list, as the comma is the
delimiter in all_truck_lists within config.ini