Rewrite choose_cabins()
The choose_cabins function started neat but slowly became more and more
bloated as more and more use cases were added. I've therefore rewritten
it so everything is shared (instead of chunks of code being duplicated,
and weaving in and out of each other with odd if statements) and the
internal_name doesn't need to be changed if editing a truck.

Also added and changed many files as part of testing a truck list and a
DAF paintjob.