Add ATS trailer database information
The placeholder entries in database.ini have been replaced by the two
actual trailers found in the game. It appears as though all parts of a
trailer are accessories, so accessories.ini has been updated as well.

Trailer accessories are a bit cumbersome, so I've temporarily grouped
them just to ensure the program's functionality, but I definitely want
to implement a better solution in the future, possibly with different
groups of accessories (so you could assign all side guards at once, or
each individually, for example).