Update paintjob.py to work with trailers
Trailer paintjobs are similar to new_truck_format truck paintjobs, but
without a "main" paintjob image, and with everything moved to a
different folder. I've therefore updated all of the common mod-creating
functions to accommodate this, and introduced the trailer variable.
man.py and auto.py were also updated to supply this new variable.