Archive older (newer?) versions
Two new versions of Paintjob Packer have been created since the end of
updates to the repo, which have been archived in a new older versions
folder along with the latest version from the repo. v1 is a copy of
b251843. v2 is a total rewrite (the files of which were originally in
the main directory with the other files, hence the cm_ prefix) making
use of an input text file, rather than a CLI. v3 is a tidied up version
of v2 that features a description generator. It also makes paintjobs in
a new format, utilising the settings.sui files seen in newer DLCs and
recent updates from SCS.

v3 of the program, which I'm actively using today, is getting too fiddly
to handle, so another total rewrite is in order. I figured I'd put it on
the repo this time so that it can be used by the public if it turns out
okay.