Copy only cabin_a.dds
Having separate .dds files for each cabin from the beginning would add
more busywork to the creation of each paintjob, as different cabin sizes
often use the same texture, and all of the .tobj files would need to be
redirected to a single .dds file. It would also negate the idea of the
unifier, as it will rely on the presence (or absence) of cabin_* .dds
files when determining which paintjobs to merge.