Link every accessory .tobj to shared_colour
Most accessories need to be linked to a paintjob's shared_colour .dds,
and the few that don't are usually linked to a second shared colour. For
that reason,, linking all of the accessories this way from the beginning
will cut down on time otherwise spent deleting .dds files and relinking
each .tobj.