Convert functions to use objects
Functions now accept objects instead of loose variables, except in the
case of cabins - they're only two variables each and probably don't need
a class of their own.

In many cases the same information is passed to a variable twice, for
example in make_settings_sui, each vehicle is already a part of the
paintjob that is given. It would be cleaner to simply iterate through
each vehicle, and have the paintjob be the only input. The issue with
this is later, when partial mods are begin created (such as when making
an addon, or running the unifier), only certain vehicles are needed.
It's thus better to keep the inputs separate, so they can be controlled
accordingly.