Separate making vehicle files
All vehicle files are now made in make_vehicle_files, rather than
make_pack, so that addons can make use of them. A new int_name was added
to related packs, also for the sake of addons. It's not needed for other
functions, namely make_workshop_description, but it allows the creation
of each related pack's .ini section when saving a pack file after an
addon has been made.