Re-restructure existing packs
The input file (and subsequently the .py file) for existing packs has
been stripped down with much of its functionality removed. I looked at
the ways I'd actually use this file, and realised that I was making a
lot more work for myself just to implement a few extra functions I
probably wouldn't ever end up using. For that reason it now simply
generates descriptions and makes addons. If I want to regenerate whole
packs I can add that easily with a make_pack, and in the unlikely event
that I need to do anything else with existing packs... I'll cross that
bridge when I come to it.