Rebuild and save modified packs
Packs are now turned back into ConfigParser objects and saved as .ini
files after addons have been created. This is technically not required,
however it makes the simultaneous creation of descriptions much easier.
In the event that a single mod gets more than one addon, it would be
impossible to keep track of all of the vehicles for the purposes of
descriptions without this (admittedly messy) function.