Unify identical paintjobs
By default each cabin of each truck gets its own paintjob (unless
separate_paintjobs is false), so that designs can be modified to work
on all sizes of truck. With many designs, however, many or all of the
different cabins can be served with a single texture, and the separate
paintjobs just make upgrading/modifying trucks more difficult. The new
unifier takes care of that, combining paintjobs based on which .dds
files their .tobjs point to.

Three new outputs have also been added to save_new_pack_to_database, as
in the case that a pack already exists, the lack of feedback was an
issue.