Copy unifier.py, and make the .ini work
The unifer.py is now copied to the output folder when the .ini is made.
There were issues with the reduction percentage message, so an extra
variable was made to make things a bit clearer.

There was also an issuewhen making the unifier.ini, as some lines were
expecting pj to be a string, rather than a Paintjob object. They've
therefore been changed to use pj.int_name.

The unifier folder and its two empty files have also been deleted as
they're no longer needed