Check for empty accessories
As accessories are now split into two different lists, there was an
issue involving any vehicle with only shared_colour accessories and no
textured accessories, or vice versa. One of their accessory lists would
be an empty string, and this was attempted to be used as a dictionary
key. This has been avoided with a quick check to make sure each
accessory isn't an empty string. This wasn't an issue initially, as any
vehicle with uses_accessories as true would have at least one accessory
in their list. The problem came about from the addition of the second
list, as it may be (and often is) empty.