Confirm that process completed successfully
A new console output of "Success!" has been added wherever the program
didn't output anything before, and the program now sleeps for 2 seconds
after any operation to give the user time to read whatever was output to
the console. This is to provide clearer feedback that everything worked,
so in the case anything goes wrong it's more obvious.

The SCS Food Tank trailer from the ATS 1.37 beta was also added to the
vehicle database.