Update pack version if adding vehicle
If adding a vehicle to an existing pack, the user is given the chance to
update the version number of the pack. Rather than trying to work it out
automatically, the current version number is displayed and the user
(optionally) enters the new one. The pack's database entry is then
updated, and a new manifest.sii is generated so that the change can be
reflected in the in-game mod manager.