Implement GUI and setup screen
Work on a GUI has now begun using tkinter, beginning with a setup screen
that's shown before the main program. goals and tkinter reference.txt
has a list of the goals of this ui overhaul, as well as a hierarchical
map of the tkinter widgets involved, and a tkinter reference cheat
sheet, as this is my first time using tkinter.