Change when OS is detected and cursor chosen
Previously the cursor was set to "hand2" when the program started up,
and only changed to the platform-specific cursor when loading the main
screen. Since the feedback link is gridded to the main screen before it
is displayed, it ignored the now-correct cursor value and always used
"hand2". By choosing the correct cursor from the get-go, the link now
uses the appropriate one. This issue only appears when running on macOS.