Add link to page with mod links
Created a list of all of the mods supported by Paintjob Packer, along
with links to download them. A link to this page has been added to the
Truck Mods tab of the vehicle selector. Hopefully this will avoid users
getting mods confused, and accidentally generating paintjobs for the
wrong mods, e.g. RTA's Argosy instead of Harven's.

At the time of implementing this link, it's pointing to a file on GitHub
that does not yet exist. After this commit is pushed, it still won't be
in the correct branch yet, which is... interesting.