Replace shared colour with unifier option
Shared colour files will not be used in this version of Paintjob Packer,
and I've replaced the option with a choice to use the cabin unifier
system. Since this system is fiddly and complex, I've provided ample
warnings for users who enable it, so they hopefully watch the
video and don't get lost when using it.

I've also changed all hyperlinks to variables, as they're used in
multiple places and may need to be changed in the future. The mod link
page link (that's a mouthful) has also been moved to its current,
temporary location in the ui-overhaul branch.