Remove textured accessories
The concept of textured accessories has been removed, with the
corresponding fields in Vehicles removed, and the entries in all vehicle
database inis. Paintjob.py has been duplicated as it will likely undergo
extensive rewriting to achieve compatibility with the new GUI, and thus
I can use the old version to reference and compare against.