Change output folder handling
Instead of outputting everything to /output, files are now created on
the desktop via an output_path variable. Loose file copying/creation
has also been rewritten.

Going forward, I'm also replacing all %s string replacement with {}
and .format, as it seems to be more modern and the preferred method
in the community.