Rewrite make_unifier_ini
The unifier.ini is much simpler since there's only ever one paintjob at
a time, and as such the rewritten make_unifier_ini is much shorter and
cleaner. The rest of the functions beneath it have also been deleted, as
they all refer to the pack library system, which has been removed, and
calling all of the required functions to make a working mod, which will
be moved to packer.py.