Generate readme and rewrite unifier.py
A readme-type text file is now generated alongside a mod, listing all of
the placeholder files that need to be replaced. The cabin unifier has
also been rewritten to work with the new format of unifier.ini files.
Special consideration has been made to ensuring all files that need to
exist to exist, and are not incorrectly named. That way, if something
goes wrong, the user should know exactly what they need to fix.

Some old entries in .gitignore have also been removed.