Replace separate & combined pictures, and add logo
The images for separate and combined paintjobs have been replaced with
new versions that more clearly show the three separate trucks. A new
logo has been designed, and has been added to the welcome screen image,
and the program icon.

All references to the instructional video should now instead refer to
the guide that will be on the GitHub page (unless I've missed one), and
the warning that shows up upon selecting the cabin unifier has been
toned way down.