Change directory if packaged, and add guide
The directory is now changed so that files (images, placeholder files,
vehicle .inis etc) work when Paintjob Packer is packed as a single file
using PyInstaller. In the event that changing the path fails, the
program is running as a .py and will instead use its location.

A guide has also been added to the readme, detailing the process of
making a paintjob mod from scratch, as it doesn't end with clicking the
generate button in Paintjob Packer.

References to Python 3 have also been removed, as the unifier works as
intended when packaged as an executable, and thus will be.