Combine error messages, and change output path
If multiple things are wrong when attempting to generate a mod, all of
the errors will now be put into a single popup. Previously they were
all shown separately as I though the only alternative was to only show
one of them, and I didn't want users to have to click Generate mod 5+
times, only being able to fix one error each time. By combining them,
the user isn't bombarded with spam, but is still given all necessary
information. If the output folder isn't clear, that error is separate
from the others due to its length and because it requires external
action.

The output path has also changed. Files are created on the desktop if
Paintjob Packer is running as a loose .py, but next to the executable
if packaged. The only reason I didn't want to output in the Paintjob
Packer folder itself is so that users don't see all of the messy files,
but when it's packaged as a single executable they don't, so that folder
is fine to use. Most of the functions in paintjob.py had to be updated
to take this extra input, as well.

The path to the appropriate mod folder has also been added to the
generated readme file.