Turn cabin unifier into tkinter program
The cabin unifier has been rewritten as a (GUI-less) tkinter program,
so that users never have to see a terminal window. As the only UI for
the unifier was warnings and prompts, the tkinter version consists of
only message boxes, with a withdrawn root window.

A few missing new line characters were also added to the generated
readme.