Add ETS 2 trailer mods, update mod links
Support for all of the ETS 2 trailer mods has been added. Some mods had
to be removed as they either crashed on startup, didn't allow custom
paintjobs, or in the case of Schumi's packs were locked so that
accessory information couldn't be found. Names were also changed based
on how the trailers appear in-game, and the links page was updated
accordingly.

The Project 3XX also had "Peterbilt" prepended to it, as it is present
on the SCS Forums page but not on the Workshop, and is a clearer
description of the mod.