Generate a separate workshop uploading readme
A second readme is now generated when uploading to Steam Workshop, with
instructions on downloading the Workshop Uploader and preparing a mod
for upload. It will also contain an automatically generated description,
but that hasn't been implemented yet.

A few typos were also corrected, including some ouput > output, and
closing the file when generating the main readme.