Generate workshop description
A Steam Workshop description is now generated inside the Workshop
uploading readme. It is identical to the mod manager description, but
has BBCode-formatted links to any supported mods, which appear as
clickable links on Steam Workshop.

Fixed a typo that meant that the "Trailers supported" section of
the description wouldn't show up sometimes.

Incremented version number, and switched forum link to the actual link,
not the shortened version.