Remove cabin unifier
The cabin unifier was a cool idea, but it ultimately added a lot of
complexity to the mod making process for no real benefits. Asking users
to dive into a TOBJ is a bad idea, and the whole process of adding
textures is rife with potential errors.

The actual benefits from using the system are not important, as the
ability to switch cabins without needing to rebuy a paintjob is not a
serious concern. Any users with the understanding and skills required
to use the unifier will still be able to take advantage of the smaller
mod size by editing the TOBJs themselves, but I no longer need to rely
on them being able to follow complex instructions by requiring them to
do so.

This will also allow me to move away from the snake case cabin file
naming in exchange for the more user friendly naming scheme. Each
cabin DDS will now contain the cabin's actual in-game name, which
would be nearly impossible to get users to replicate when adding a new
texture in the old unifier system.