Change folder structure in vehicle folder
All of the TOBJs and DDSs have been moved from vehicle/<type>/upgrade/
paintjob/<internal name>/<make>_<model>/ to vehicle/<type>/upgrade/
paintjob/<in-game name>/<vehicle name>/. This way users are never
exposed to internal/ugly names when making their mods. As of now, the
only snake case name they see in the mod creation process is the MAT
file for the paintjob icon.

The mod manager description and image names have also been changed for
the same reason, although since the image name cannot have spaces in it,
they've been renamed to a single word each for consistency.

The Volvo FH16 2009 was also updated to the new cabin/accessory naming
scheme, as part of testing.

I recognise that this change, as well as the change to the DDS naming
scheme goes against the conventions of the games and most mods. In this
case, however, I'm happy breaking convention in the name of user
friendliness and accessibility - I want Paintjob Packer to be as easy to
use as possible, in order to introduce more people to the world of
trucksim modding. More experienced modders may be frustrated with this
bucking of tradition, but I feel Paintjob Packer is made with the more
casual modder in mind, especially given the changes in version 1.4.