Temporarily remove modded vehicles
All truck and trailer mods have been moved to a subfolder to keep them
out of the way while the vanilla vehicles are transitioned to the new
cabin/accessory system. They'll be transitioned after the kinks have
been ironed out with just the vanilla vehicles. There are also no
templates for them yet, as I need to investigate whether or not adding
(all) modded vehicles to my template pack is feasible.