Don't add display_name if uploading to Workshop
Whenever a mod is uploaded using the SCS Workshop Uploader, a warning is
always given regarding the display_name property in the manifest.sii.
This warning can be safely ignored (I always ignore it), but it can be
off-putting for a new modder. The display_name property is therefore no
longer included in the manifest.sii if the Workshop uploading box is
ticked, and I've confirmed in testing that the mod works fine, still
validates, and the warning no longer appears. In the event that a user
chooses this option and then does not upload their mod to the Workshop,
the mod will use the folder name in the mod manager, which is also the
mod's name by default, so there're no worries there.