Add placeholder templates option
A checkbox has been added to the internal settings panel to use the new
placeholder templates. This option is disabled if the templates are not
installed, and does not affect modded trucks or trailers. If disabled,
the empty.dds file will be copied and renamed, as in v1.3

A number of unused files have also been removed, and the version number
has been incremented.