Move modded ETS 2 vehicles out of temporary folder
All of the 12 modded vehicles that have been transitioned to the new
cabin and accessory system have been moved out of the "mods" subfolder
so that their paintjobs can be tested. An additional 3 vehicles have
also been transitioned, and the Mercedes-Benz Antos has been removed as
it caused crashes.

The 15 modded vehicles have also had (empty) placeholder files created,
and I've completed their parts of my template pack. Once all of the ETS
2 vehicles have been transitioned, and the new ones added, these
templates will be added to the placeholder files, although the changes
will not appear as they are covered by the .gitignore. The vehicle
database files will likely change as I understand the textures more, for
example I've changed the Schmitz Cargobull and Menci Cistern to
non-accessory vehicles as all of the accessory textures are UV mapped
specifically to work on a single vehicle texture. I'll have to update
how paintjobs are made to account for this.

The lines trying to load the images removed on a whim in 135d3e5 have
also been removed, as they caused crashes on startup.