Transition all ATS mods to the new system
Each ATS modded truck and trailer has been switched to the new accessory
and cabin system, and has had its make and model removed. Through the
course of making templates for each of them, a couple have also been
modified, namely all of the trailers being switched to non-accessory
vehicles, and some rearrangement of cabins on a few trucks (such as
the Scot A2HD, with a nonexistent sleeper cabin removed).