Add new generating screen
A new screen has been added that appears after clicking the Generate
button on the main screen. A few errant options have been moved to it
(workshop uploading and placeholder templates), and it gives users the
opportunity to see their save directory before saving. The progress bar
has also been moved to it.

The final Generate button is not yet functional, nor is the progress
bar.

A typo was also corrected in the International 9800 cabins.