Check for alt uvset when copying template
When extracting a single main DDS for a truck with multiple cabins,
Cabin A's name is read from the vehicle's database file and used to
target the correct .DDS. For vehicles using the alt uvset, this name
is appended with "alt uvset" to make it clear to the user. This
appendment is done in make_paintjob by modifying main_dds_name, but
since copy_main_dds read directly from the vehicle database file this
step was skipped. I've corrected this by shoehorning in a little check
that appends it within main_dds_name instead.