Rewrite readme guide
The guide on the GitHub readme has been updated to reflect the changes
in v1.4, including the addition of placeholder templates and the removal
of the cabin unifier. Links to my template packs have also been added
for convenience.

A few help buttons' prompts were also updated to clean up their wording,
and update_cabin_dropdowns is now called when changing the cabin
handling dropdown to ensure the maximum internal name length is always
updated correctly.