Check for library folder on startup
Because some users will be running Paintjob Packer from the .py file,
it's important to check that the library folder and paintjob.py are
present and display an error if not, rather than just closing. Importing
modules happens before tkinter starts, so show_unhandled_error doesn't
get a chance to start. Everything else that happens until it gets that
chance *should* just rely on preinstalled Python modules, and not be an
issue. I hope.