Improve placeholder images
The paintjob icon has been resaved with premultiplied alphas, so it no
longer shows up with white crust around it in-game. All placeholder
images have also been modified to remove the big Paintjob Packer logos,
so that the important information (namely image size) is more prominent.
As premultiplying the alphas of the icon removed the crust, there is no
longer a need to fiddle with the TOBJs, so TODOs in paintjob.py
referring to it have been removed.