Move version number to separate file
The current version of the program is now stored in a separate text
file named version.txt, to allow for update checking. The current
version can be compared to the latest version on GitHub (by downloading
the raw text file), to determine whether or not to show the new version
notice to the user.