Copy unhandled exception to clipboard
In the event of an unexpected error, the full error message and
traceback is now copied to the user's clipboard instead of just being
shown in an error message. Error message text can't be selected and
copied, so this way an error can actually be submitted in case anything
ever goes wrong. The user is also now not shown a giant wall of text in
an error window, which they'd likely ignore.