Move vehicle checkboxes in a scrolling list
When making a paintjob pack, vehicle selection checkboxes are now
arranged in a scrolling list, rather than split up into two columns.
This was always how I wanted them to be displayed, but at the beginning
of the project I was unable to find a tutorial that explained how to
make scrolling frames (I evidently didn't look very hard), so I passed
it off as too complex. It's important that I implement it now, as with
mod authors displayed in the list, it was getting very wide.