Add new version info INI file
The current version checking system works, but it's really not good.
This is the beginning of a test to see if I can switch to a new
system powered by configparser, which will allow me to display a
short message when an update is available, informing the user about
what they'll gain from updating. If all goes accoring to plan,
version.txt will no longer be used, but will be maintained and updated
for a decent while to allow version 1.7+ to propagate throughout the
userbase.