Rename version info to version.ini
urllib.request.urlopen() can't parse a URL with a space in it, so the
file has been renamed to version.ini. I could have converted it to kebab
case or snake case, but in keeping consistent with the rest of the files
in this project (e.g. mod links.md and the vehicle database entries)
it's easier to simply make it a single word.