Add paintjob tracker checklist stage
A number has been added which represents which stage of the checklist a
pack is up to. The changelog additions to existing packs have also been
rewritten to reflect the move from a section to a single key.

If an update is not detected, a message has been added to the start
screen warning that things might look a little weird on macOS or Linux.