Update mod links.md
Changed links for most of XBS' mods to their website, as they are releasing updates on there, rather than the SCS forums. The main exception for ETS2 is the Volvo F88, which is not on the website, so the link has been left as the SCS forums. For ATS, only the Freightliner FLC is on their site, so that has been changed too.