Extend DBus warning to Cristhian's bus mod
As well as Etrusan's mods, the warning explaining that paintjobs don't
work properly is now also displayed for Cristhian's Caio Millennium 2
PBC, and I suspect it will also need to be displayed for any other bus
mods in the future.

The formatting of Workshop descriptions was also tweaked to prepend each
item in a list with a dash, to aid readability. Mods named "123" also no
longer trigger Paintjob Tracker, as that is my go-to mod name for making
a quick test mod.