Fix VAK trailer body names
The body names of the VAK trailer are inconsistent, 4_4_curtain is
alternately called Curtain, Curtainside and Tarpaulin Roof in the body
def and the two body configurations available. Due to the confusion, the
templates were incorrectly named, and didn't work. Since the names are a
little confused and there doesn't seem to be a single answer, I've
renamed it Curtainsider in both the vehicle database file and templates
to match the SCS Box Trailer, and renamed 4_4_dry and 4_4_reefer Dry
Freighter for the same reason.

There was also a typo on 4_4_curtain which meant it didn't work, which
has been fixed.