Remove old DAF 2021 file
Because of the new vehicle database file naming convention including a
vehicle's author, when new-gen-daf was merged into master for v1.7 the
old daf.2021 file didn't get replaced or removed, and so caused issues
when Paintjob Packer was trying to load the list of vehicles. It has
now been removed.