Check for write permission in save directory
An empty DDS is copied into the save directory and then deleted to
ensure that Paintjob Packer has write permission. If not, instead of
showing the error/crash screen, a small popup is now shown instead,
giving users the opportunity to choose a different folder or try running
Paintjob Packer as an admin.