Catch incorrectly expanded desktop folder
As reported by users, sometimes os.path.expanduser doesn't work as
expected, resulting in an invalid save path. This is causes an uncaught
FileNotFoundError. This error is now checked for when testing write
permission, and if found it sends the no permission dialogue to the
user. This means that the error message isn't 100% correct, but it does
achieve the same result, as the steps to fix it (choose a different
directory) are identical.