Move templates to own folder
To make it easier to build templates and non-templates versions of the
installer separately, the templates have been moved to their own folder,
instead of being inside the library folder. This has the upside of also
being more accessible to anyone looking through the program's files to
find the template .DDSes.