Load system default language
If the default system language is supported by Paint Job Packer, it is
loaded. If not, if a different dialect is supported, it will be loaded,
e.g. pt-PT instead of pt-BR. If none are supported, en_GB is loaded by
default, though this will change to en_US once it's ready.

The full stop is now just called the full stop (its British English
name) in en_GB, with the additional "/period" removed, as en_US will be
a separate language.