Only fetch new version info once
Since the PackerApp is restarted each time the language is changed,
version info was re-fetched as well. By making new_ver a global variable
it persists between PackerApps, and therefore only needs to be fetched
once. This likely won't have a huge effect on anything, but since I'm
not really using the GitHub APIs properly anyway, it's good to reduce my
impact on them.