Opt out of SSL validation for version checking
On macOS, Python comes with SSL certificates but does not install them
itself, maybe for security/sandboxing reasons. The user can install them
at a later date, and is prompted to do so in the installer dialogue, but
there's no way to verify that a Paint Job Packer user has done so, nor
any way to automate it to my knowledge. This causes a hidden SSL
certificate error when running on macOS (only visible in the system.log
in the Console app), meaning version fetching always fails. By creating 
an unverified SSL context and requesting the version using it, I am
effectively opting out of the default certificate verification, as
detailed in PEP 476.