Default to English when system_lang is None
locale.getdefaultlocale() returns None if the app does not have access
to environment variables, which is the case when packaged for macOS
using PyInstaller. This is now accounted for, and the program will
default to English when this occurs. This commit also changes the
default language to en_US, as it is the more widely used dialect and is
now available thanks to Crowdin.