Remove Paint Job Tracker file when building
The Paint Job Tracker file is .gitignore-d so should only exist on my
computer, but when building the programs it's included as it's in the
library folder, causing issues when running Paint Job Packer if the
specified PJT path doesn't exist. It's now removed at the start of each
build script. This shouldn't affect my own installation, as all the
files are copied to a new folder before building.