Possibly fix unescaped localisation comment
The ErrorSingle context comment was the only comment not separated from
the string by a space, and in some cases it showed up in the UI. The
space has been removed, and the comment no longer appears. This may have
been the cause of the bug, but if not, it's nonetheless a small
improvement to the file's tidiness.