Send basic analytics data
Each time Paint Job Packer is used to generate a mod, it now sends a
small packet of information to RudderStack which logs it in a
spreadsheet. This is handled in analytics.py, wherein every function and
piece of data is clearly explained so that even non-technical users know
exactly what data is being collected about them. The RudderStack
consumer has been modified to print when it has finished uploading, and
the client's timeout has been reduced from 15 seconds to 7, as waiting
15 seconds after the program is finished is a bit ridiculous.