Create misc utilities and update mod links.md
A misc utilities script has been written that automates a few Paint Job
Packer management tasks, including the creation of mod links.md. The
file has been re-generated (or generated for the first time, as it was
hand-written) with the recently added vehicles, and the updated vehicle
names including diacritics.