Update mod links and FORD F-MAX database file
mod links.md has been updated with all of the recently-added vehicles,
and the FORD F-MAX database file has been renamed to fit its author.
The misc utilities script was also updated to check for incorrectly
named/missing accessories and template textures, leading to a few
updates to the templates (which are .gitignored). The Peterbilt 352
was also removed as the templates didn't really work, and it doesn't
seem to have been updated recently.