Load the correct single paint job vehicle file
When making a single paint job instead of a paint job pack, the file
loaded was chosen only using the vehicle's display name. This is not a
problem most of the time, except for when trying to make a paint job for
the Scania R, as scania.r [RJL] comes before scania.r [SCS]
alphabetically, and they have the exact same display name. The author is
now checked as well when choosing which file to load. The author of each
mod is now also displayed in the vehicle selection dropdown.