Catch path length-related FileNotFoundError
A number of users experiences crashes caused by Windows' max path length
of 260 characters by default, these crashes are now caught and an
appropriate error message is displayed. There are ways to force the
program to write the files regardless of the path length, however afaik
this can also lead to the creation of then-undeletable files (like files
with system-reserved file names), so in the interest of not wanting to
walk a user through the complicated process of deleting such a file,
it's easier to simply request that they choose a shorter output path.