Add missing trailer parts
The UV maps of the ETS 2 SCS Box Trailer sideskirts have returned, and
so the templates and accessory support have too. Not every size of
sideskirt is placed in its own accessory group, to avoid overwhelming
the user.

The new Paint II rear bumper on Krone trailers was also added, and a fix
was applied to the Schwarzmüller Food Tank allowing its chassis and
sideskirts to be repainted, utilising the bus door workaround feature.