Fall back to time.monotonic()
Part of the RudderStack script uses monotonic.monotonic(), which is an
alias of time.monotonic(), however it seems that sometimes the monotonic
module is not included in the Python installation, resulting in an
error. Switching solely to time.monotonic() would cause issues on early
Python versions, so imports for both are attempted.