Catch crash when checking for outdated files
When checking for outdated vehicle database files, if one is found Paint
Job Packer will attempt to delete it. On at least one occasion this
resulted in a PermissionError crash. The crash still happens, as I'm not
sure how to programmatically resolve this, but the user is at least
warned beforehand, and instructed to reach out to me.

Crash tracked in issue #21, this doesn't completely fix it, but at least
improves the experience.